+++ Rules/IPv4/Outbound/RemoteWindows.ps1
@@ -133,6 +133,23 @@ properties control panel item." |
 	Format-RuleOutput
 }
 
+$Program = "%SystemRoot%\System32\wsmprovhost.exe"
+if ((Test-ExecutableFile $Program) -or $ForceLoad)
+{
+	# TODO: Unknown why wsmprovhost.exe needs internet access to MS servers
+	New-NetFirewallRule -DisplayName "PowerShell remoting" `
+		-Platform $Platform -PolicyStore $PolicyStore -Profile Private, Domain `
+		-Service Any -Program $Program -Group $Group `
+		-Enabled True -Action Block -Direction $Direction -Protocol TCP `
+		-LocalAddress Any -RemoteAddress Internet4 `
+		-LocalPort Any -RemotePort 443 `
+		-LocalUser $AdminGroupSDDL `
+		-InterfaceType $DefaultInterface `
+		-Description "Host process for WinRM plug-ins. Process wsmprovhost hosts the active remote
+session on the target. When a remote PowerShell session starts, svchost.exe executes wsmprovhost.exe" |
+	Format-RuleOutput
+}
+
 if ($UpdateGPO)
 {
 	Invoke-Process gpupdate.exe -NoNewWindow -ArgumentList "/target:computer"