+++ Rules/IPv4/Outbound/WindowsSystem.ps1
@@ -291,7 +291,7 @@ $Program = "%SystemRoot%\System32\backgroundTaskHost.exe"
 if ((Test-ExecutableFile $Program) -or $ForceLoad)
 {
 	# TODO: need to check if port 22 is OK.
-	# TODO: Dropped connection (for admin), likely reason for app downarrows
+	# TODO: Dropped connection (for admin and user), likely reason for app downarrows
 	# NOTE: Testing with profile "Any"
 	New-NetFirewallRule -DisplayName "Background task host" `
 		-Platform $Platform -PolicyStore $PolicyStore -Profile Any `